finished merging, main class is now TicTacToe.java
